-refactor directories
