+++ src/main/java/by/asckarugin/dto/hh/Area.java
@@ -1,12 +1,14 @@
-package by.asckarugin.dto;
+package by.asckarugin.dto.hh;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
-public class Type {
+public class Area {
     @JsonProperty("id")
     private String id;
     @JsonProperty("name")
     private String name;
+    @JsonProperty("url")
+    private String url;
 }
