+++ src/main/java/by/asckarugin/dto/hh/Employment.java
@@ -1,14 +1,12 @@
-package by.asckarugin.dto;
+package by.asckarugin.dto.hh;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
-public class Area {
+public class Employment {
     @JsonProperty("id")
     private String id;
     @JsonProperty("name")
     private String name;
-    @JsonProperty("url")
-    private String url;
 }
