+++ src/main/java/by/asckarugin/dto/hh/LogoUrls.java
@@ -1,4 +1,4 @@
-package by.asckarugin.dto;
+package by.asckarugin.dto.hh;
 
 import lombok.Data;
