+++ src/main/java/by/asckarugin/dto/hh/Type.java
@@ -1,10 +1,10 @@
-package by.asckarugin.dto;
+package by.asckarugin.dto.hh;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
-public class Employment {
+public class Type {
     @JsonProperty("id")
     private String id;
     @JsonProperty("name")
