+++ src/main/java/by/asckarugin/dto/hh/Vacancy.java
@@ -1,5 +1,6 @@
-package by.asckarugin.dto;
+package by.asckarugin.dto.hh;
 
+import by.asckarugin.dto.hh.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
